NXP-19435: consolidate inner helper classes
